refactor: extracted analytics error functions into helpers
